requirements.txt
Django>=4.2
djangorestframework
paho-mqtt